examples/langgraph/requirements.txt
# LangGraph Agent dependencies
langgraph>=0.2.74
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0
aiohttp>=3.8.0
tiktoken>=0.5.0